Backend - Officers Seeds - Implemented
Updated the name of the Officers model to fit conventions. Also added in total_complaints and total_substantiated_complaints to the .create() method
